+++ src/js/core/toggle.js
@@ -1,4 +1,4 @@
-import { $, hasTouch } from '../util/index';
+import { $ } from '../util/index';
 
 export default function (UIkit) {
 
@@ -21,10 +21,6 @@ export default function (UIkit) {
             mode: 'click',
             queued: true,
             media: false
-        },
-
-        init() {
-            this.mode = hasTouch && this.mode == 'hover' ? 'click' : this.mode;
         },
 
         events: [